+++ server/game/engine/CallManager.cpp
@@ -1,3 +1,5 @@
+#include "server/precompiled.hpp"
+
 #include "server/game/engine/CallManager.hpp"
 #include "server/game/engine/EntityManager.hpp"
 #include "server/game/engine/Engine.hpp"
@@ -64,7 +66,7 @@ namespace Server { namespace Game { namespace Engine {
         double seconds = helper.PopArg().CheckNumber("Server.Call.Later: Argument 1 (seconds) must be a number");
         if (seconds < 0)
             seconds = 0;
-        Uint32 targetId = helper.PopArg().CheckNumber("Server.Call.Later: Argument 2 (target) must be a number");
+        Uint32 targetId = (Uint32)helper.PopArg().CheckNumber("Server.Call.Later: Argument 2 (target) must be a number");
         std::string function = helper.PopArg().CheckString("Server.Call.Later: Argument 3 (function) must be a string");
         Tools::Lua::Ref arg(this->_engine.GetInterpreter().GetState());
         Uint32 cbTargetId = 0;
@@ -75,7 +77,7 @@ namespace Server { namespace Game { namespace Engine {
             arg = helper.PopArg();
             if (helper.GetNbArgs())
             {
-                cbTargetId = helper.PopArg().CheckNumber("Server.Call.Later: Argument 5 (cbTarget) must be a number");
+                cbTargetId = (Uint32)helper.PopArg().CheckNumber("Server.Call.Later: Argument 5 (cbTarget) must be a number");
                 cbFunction = helper.PopArg().CheckString("Server.Call.Later: Argument 6 (cbFunction) must be a string");
                 if (helper.GetNbArgs())
                     cbArg = helper.PopArg();
@@ -85,7 +87,7 @@ namespace Server { namespace Game { namespace Engine {
         Uint32 notificationCallbackId = 0;
         if (cbTargetId)
             notificationCallbackId = this->_engine.GetCallbackManager().MakeCallback(cbTargetId, cbFunction, cbArg);
-        this->_calls[this->_engine.GetCurrentTime() + seconds * 1000000.0].push_back(new Call(targetId, callbackId, notificationCallbackId));
+        this->_calls[this->_engine.GetCurrentTime() + (Uint64)(seconds * 1000000.0)].push_back(new Call(targetId, callbackId, notificationCallbackId));
     }
 
     void CallManager::_ApiNow(Tools::Lua::CallHelper& helper)
